refactor: strcuctre
