--- data-source/go.mod
@@ -1,8 +1,0 @@
-module github.com/yomorun/yomo-source-example
-
-go 1.15
-
-require (
-	github.com/yomorun/y3-codec-golang v1.6.9
-	github.com/yomorun/yomo v1.2.520
-)
